+++ src/main/java/com/apisql/ApiSQL/repository/TipoRiscoRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoRiscoRepository extends JpaRepository<TipoRisco, Long> {
+public interface TipoRiscoRepository extends JpaRepository<TipoRisco, Integer> {
 }
